+++ CHANGELOG.md
@@ -3,6 +3,7 @@
   - Fixed -ExecutionPolicy in pester.bat [GH-130]
   - Add support for mocking internal module functions, aliases, exe and filters. [GH-126]
   - Fix TestDrive clean up [GH-129]
+  - Fix ShouldArgs in Strict-Mode [GH-134]
 
 ## 2.0.4 (March 9, 2014)